refactor: 💡 now automatically FGA_STORE_ID value will be added
i have added a function it will do this, check in .env FGA_STORE_ID is
nonempty or not assigned or not found. if found and has value then it
will leave it. but if not available or value is not there then it will
add this in .env nv variable